+++ src/app/office-view/delete-staff/delete-staff.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { StaffService } from 'src/app/services/staff.service';
+import { Location } from '@angular/common';
 
 @Component({
   selector: 'app-delete-staff',
@@ -12,7 +13,8 @@ export class DeleteStaffComponent implements OnInit {
   // Declaring variables
   staffId = "";
 
-  constructor(private activeRoute: ActivatedRoute ,private staffService: StaffService, private router: Router) { }
+  constructor(private activeRoute: ActivatedRoute ,private staffService: StaffService, 
+    private location: Location) { }
 
   ngOnInit(): void {
     this.activeRoute.params.subscribe(data => {
@@ -20,11 +22,16 @@ export class DeleteStaffComponent implements OnInit {
       console.log(this.staffId)
     });
 
-    this.staffService.deleteStaff(this.staffId).subscribe(data => {
-      console.log('Staff Deleted!');  // Deletes staff with specific id
-    });
+    let choice = confirm("Are you sure you want to remove?")
 
-    this.router.navigateByUrl('home');
+    if (choice === true) {
+      this.staffService.deleteStaff(this.staffId).subscribe(data => {
+        console.log('Staff Deleted!');  // Deletes staff with specific id
+        this.location.back()
+      });
+    }else {
+      this.location.back();
+    }
   }
 
 }
